asadiqbal08: Generalize ModuleSystem 'service' method
